Introduce `recv_blocking()` method on `PayloadReceiver`
Allows callers to wait for a frame without using futures.

Also bump async-std dependency to 1.12 (unfortunately still doesn't ensure recent-enough
`async-channel` - but I'll fix that in another PR).

Also update the stream example and READMEs with the new method, making it print any receive errors
(but not panic on them) and doing exactly 10 tries (instead of waiting for 10 success payloads).