Use std::mpsc rendezvous channel instead of pair of futures::channels
The only use of the (cancellation, completion) channel pair is to synchronize teardown of streaming
(either explicitly or when StreamHandle is dropped). This is also the only place where we were
spinning and async runtime.

Turns out `std::sync::mpsc::sync_channel(0)` is as special "rendezvous" mode that should do what we
want: sends a message and waits for it to be picked up. We also continue to detect StreamHandle drop
from the streaming loop.

Should be functionally equivalent, just with less dependencies and without async runtime.

Needs testing.